Added include guards to the headers and added missing 'extern' to variable declarations in header files.
Deleted global scope `threads` variable definition from soft_pwm.c and event_gpio.c. The thread ID that would have been stored in the variable is never accessed. Resorting to a stack variable instead. The variable address should not be stored elsewhere as the thread is referred by the ID stored in the variable by pthread_create, which is never used. If the ID is used in future, it should be added to the struct associated with the pin.

This is likely an issue with the original Python code as well.